pyproject.toml
[project]
name = "garmin-ble-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "websocket-client>=1.9.0",
]
